refactor:修改GameSrv减少不必要的分配(包裹物品)

+++ src/GameSrv/GameCommand/Commands/BackStepCommand.cs
@@ -29,7 +29,7 @@ namespace GameSrv.GameCommand.Commands
             }
             else
             {
-                playObject.CharPushed(M2Share.RandomNumber.RandomByte((byte)nType), nCount);
+                playObject.CharPushed(M2Share.RandomNumber.RandomByte(nType), nCount);
             }
         }
     }
